chore(feedback) implement reducer tests
- reducer tests are now available
- front end tests are now broken down into components
- Admin can now categorise books by categories on creation

[delivers #154131519]
