+++ server/helpers/JwTokens.js
@@ -4,14 +4,10 @@ import crypto from 'crypto';
 class JwTokens {
   // random 24 character string
   static randomString(modifier) {
-    return new Promise((resolve, reject) => {
+    return new Promise((resolve) => {
       crypto
         .randomBytes(modifier, (error, buf) => { // generate random bytes
-          if (error) {
-            reject(error);
-          } else {
-            resolve(buf.toString('hex'));
-          }
+          resolve(buf.toString('hex'));
         });
     });
   }
